added code to TeleopDrive to do minute turns, changed LaunchBall and GrabBall commands to linearly decrease motor power, added Autonomous mode selection to SmartDashboard

+++ src/org/northwestrobotics/frc2014/commands/autonomous/DirectionalCommand.java
@@ -39,7 +39,7 @@ import org.northwestrobotics.frc2014.commands.TimedCommand;
 public class DirectionalCommand extends CommandBase {
     private final int direction;
     
-    public DirectionalCommand(int direction, int time) {
+    public DirectionalCommand(int direction, double time) {
         setTimeout(time);
         requires(drivetrain);
